A simple placeholder health bar for the player
Also some commenting and modifications to other scripts to make future work easier.

+++ Assets/Code/Health and Damage/IDamageable.cs
@@ -4,8 +4,9 @@ using UnityEngine;
 
 namespace GamecampPeli
 {
+    // Interface with a DealDamage function that is used for damaging enemies and the player
     public interface IDamageable
     {
-        void DealDamage(float damage);
+        void DealDamage(int damage);
     }
 }
